25.05.12

+++ Class 2/11050.py
@@ -2,4 +2,4 @@ import math
 
 N, K = map(int, input().split())
 
-print(math.factorial(N) / (math.factorial(K) * math.factorial(N-K)))
+print(int(math.factorial(N) / (math.factorial(K) * math.factorial(N-K))))
